add(manage permission): completed user's permission

+++ app/Http/Controllers/Backend/LanguageController.php
@@ -9,6 +9,7 @@ use App\Models\Language;
 use App\Services\LanguageService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\App;
+use Illuminate\Support\Facades\Gate;
 
 class LanguageController extends Controller
 {
@@ -22,6 +23,7 @@ class LanguageController extends Controller
 
     public function index(Request $request)
     {
+        Gate::authorize('modules', 'language.index');
         $languages = $this->languageService->paginate($request);
         return view('Backend.language.index', [
             'languages' => $languages
@@ -30,11 +32,13 @@ class LanguageController extends Controller
 
     public function create()
     {
+        Gate::authorize('modules', 'language.create');
         return view('Backend.language.create');
     }
 
     public function store(StoreLanguageRequest $request)
     {
+        Gate::authorize('modules', 'language.create');
         if ($this->languageService->create($request)) {
             return redirect()->route('language.index')->with('success', __('alert.addSuccess', ['attribute'=> __('dashboard.language')]));
         }
@@ -43,6 +47,7 @@ class LanguageController extends Controller
 
     public function edit(Language $language)
     {
+        Gate::authorize('modules', 'language.update');
         return view('backend.language.create', [
             'language' => $language
         ]);
@@ -50,7 +55,7 @@ class LanguageController extends Controller
 
     public function update($id, UpdateLanguageRequest $request)
     {
-
+        Gate::authorize('modules', 'language.update');
         if ($this->languageService->update($id, $request)) {
             return redirect()->route('language.index')->with('success', __('alert.updateSuccess', ['attribute'=> __('dashboard.language')]));
         }
@@ -60,6 +65,7 @@ class LanguageController extends Controller
 
     public function delete(Language $language)
     {
+        Gate::authorize('modules', 'language.delete');
         return view('backend.language.delete', [
             'language' => $language
         ]);
@@ -67,6 +73,7 @@ class LanguageController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('modules', 'language.delete');
         if ($this->languageService->destroy($id)) {
             return redirect()->route('language.index')->with('success', __('alert.deleteSuccess', ['attribute'=> __('dashboard.language')]));
         }
